Fixed up calculate variable to be a long int

+++ application/Source/application.c
@@ -19,7 +19,8 @@ extern UART_HandleTypeDef huart2;
 
 void Application_RunDefaultTask(void)
 {
-	int16_t calculate = 0;
+	//int16_t calculate = 0;
+	long int calculate = 0;
 
 	printf("image_id: %d, version: %d.%d.%d-%s\n", (int) IMAGE_ID, (int) VERSION_MAJOR, (int) VERSION_MINOR, (int) VERSION_BUGFIX, SHORT_GIT_HASH_STRING);
